+++ GeneralsMD/Code/GameEngine/Source/GameNetwork/Network.cpp
@@ -80,7 +80,7 @@ struct ConnectionMessage
 {
 	Int id;
 	NetMessageFlags flags;
-	UnsignedByte data[MAX_MESSAGE_LEN];
+	UnsignedByte data[MAX_PACKET_SIZE];
 	time_t lastSendTime;
 	Int retries;
 	Int length;